add cinn_to_pd rules for generate_shape op
